componentizacao de cards
