+++ src/pages/FeedPage/feedPage.style.js
@@ -22,21 +22,3 @@ export const ContainerTiposComida = styled.div`
 export const TiposDeComida = styled.p`
   margin-right: 20px;
 `;
-
-export const CardContentStyle = styled(CardContent)`
-  border: solid 1px gray;
-  height: 60px;
-`;
-
-export const CardMediaStyle = styled(CardMedia)`
-  && {
-    margin-top: 10px;
-  }
-`;
-export const NameRestaurant = styled(Typography)`
-  color: rgb(232, 110, 90);
-`;
-export const ContainerInfos = styled.div`
-  display: flex;
-  justify-content: space-between;
-`;
